Simplify and speed up y4m decoding
This moves an assert out of a hot loop,
which provides some speed up in decoding y4m video,
although the impact on the overall metric times is negligible,
since the y4m decoding is not responsible for
a large portion of time consumed.

It also removes an unnecessary function parameter,
simplifying the code.