av_metrics: Show MSE and RMSE
MSE and RMSE are the classic metrics used for analysis, while modern
objective metrics outperform MSE and RMSE, for fast computation these
will be useful.

This may not be the best way to get the results. Ideally, we need to
figure a way to do partial caching of results of PsnrResults and then
use those values for MSE and RMSE.